+++ examples/reports/advanced/input_and_export/InputAndExport.php
@@ -16,7 +16,7 @@ class InputAndExport extends \koolreport\KoolReport
     function defaultParamValues()
     {
         return array(
-            "customerNumber"=>null,
+            "customerNumber"=>0,
         );
     }