+++ examples/reports/chartjs/intro/MyReport.php
@@ -1,0 +1,6 @@
+<?php
+
+class MyReport extends \koolreport\KoolReport
+{
+
+}