+++ examples/reports/chartjs/intro/index.php
@@ -1,0 +1,7 @@
+<?php
+
+require_once "../../../../koolreport/autoload.php";
+require_once "MyReport.php";
+
+$report = new MyReport;
+$report->run()->render();