+++ examples/reports/pivot/pivot_extract/index.php
@@ -1,0 +1,5 @@
+<?php
+require_once "SalesPivotExtract.php";
+$salesPivotExtract = new SalesPivotExtract;
+$salesPivotExtract->run()->render();
+?>