+++ koolreport/CHANGELOG.md
@@ -1,5 +1,12 @@
+# Version 2.33.1
+    1. Adding `AccumulativeColumn` to generate accumulative column.
+    2. Fix Group process for not counting records correctly.
+    3. Enhance the `autoload` of KoolReport
+    4. A bundle of small fixes
+
 # Version 2.32.8
     1. Revert back to previous parambinding of PDODataSource
+
 
 # Version 2.31.7
     1. Fix the bug of incorrect active report when subReport is used.